showdown with some mocked parts

+++ sapphire/contracts/ActualPokerHandProvider.sol
@@ -19,13 +19,13 @@ contract ActualPokerHandProvider is PokerHandProvider
     // secretly store the community cards and the player cards, 
     // and broadcast the encrypted player cards 
     function populate_cards(uint table_id, uint handNum, address[] memory players) internal {
-        uint[] memory cards = generateCards(players.length*2+5);
-        for (uint player_index = 0; player_index < players.length; player_index++) {
-            uint hole_1_index = player_index * 2; 
-            uint hole_2_index = player_index * 2 + 1; 
+        uint8[] memory cards = generateCards(uint8(players.length*2+5));
+        for (uint8 player_index = 0; player_index < players.length; player_index++) {
+            uint8 hole_1_index = player_index * 2; 
+            uint8 hole_2_index = player_index * 2 + 1; 
 
-            uint hole_1_card = uint8(cards[hole_1_index]);
-            uint hole_2_card = uint8(cards[hole_2_index]);
+            uint8 hole_1_card = uint8(cards[hole_1_index]);
+            uint8 hole_2_card = uint8(cards[hole_2_index]);
 
             // store the cards secretly on chain
             playerCards[players[player_index]][table_id][handNum] = PokerHandValidation.PlayerCards(hole_1_card, hole_1_card);
@@ -56,15 +56,15 @@ contract ActualPokerHandProvider is PokerHandProvider
         );
     }
 
-    function generateCards(uint howMany) private view returns (uint[] memory) {
-        uint[] memory toReturn = new uint[](howMany);
-        uint count = 0;
+    function generateCards(uint8 howMany) private view returns (uint8[] memory) {
+        uint8[] memory toReturn = new uint8[](howMany);
+        uint8 count = 0;
         
         // TODO - THIS WHILE LOOP IS A LITTLE UGLY. 
         // BUT GENERATING AND REJECTING WAS EASY AND OBVIOUSLY CORRECT
         // GOT A BETTER WAY? LET'S DO IT
         while (count < howMany) {
-            uint candidate = uint8(Sapphire.randomBytes(1, "")[0]);
+            uint8 candidate = uint8(Sapphire.randomBytes(1, "")[0]);
             if (candidate < 52 && !contains(toReturn, candidate, count)) {
                 toReturn[count] = candidate;
                 count++;
@@ -74,7 +74,7 @@ contract ActualPokerHandProvider is PokerHandProvider
         return toReturn;
     }
 
-    function contains(uint[] memory array, uint value, uint arrayLength) private pure returns (bool) {
+    function contains(uint8[] memory array, uint8 value, uint8 arrayLength) private pure returns (bool) {
         for (uint i = 0; i < arrayLength; i++) {
             if (array[i] == value) {
                 return true;
